+++ CHANGELOG.md
@@ -3,3 +3,14 @@
 ## [1.0.4] - 2023-12-21
 
 - **Release package**
+
+## [1.0.5] - 2023-12-22
+
+### Fix:
+
+- Remove the custom text option
+- Remove size option in math
+
+### Add:
+
+- Document package